Revert "Add a locator for executables on $PATH on Windows. (#14675) (#14951)"
This reverts commit 651a6b94f790f61ca09c9c87bb712f3f5d930b9a.
