+++ src/client/pythonEnvironments/common/externalDependencies.ts
@@ -7,7 +7,6 @@ import * as vscode from 'vscode';
 import { ExecutionResult, IProcessServiceFactory, SpawnOptions } from '../../common/process/types';
 import { IExperimentService } from '../../common/types';
 import { chain, iterable } from '../../common/utils/async';
-import { normalizeFilename } from '../../common/utils/filesystem';
 import { getOSType, OSType } from '../../common/utils/platform';
 import { IDisposable } from '../../common/utils/resourceLifecycle';
 import { IServiceContainer } from '../../ioc/types';
@@ -16,8 +15,6 @@ let internalServiceContainer: IServiceContainer;
 export function initializeExternalDependencies(serviceContainer: IServiceContainer): void {
     internalServiceContainer = serviceContainer;
 }
-
-// processes
 
 function getProcessFactory(): IProcessServiceFactory {
     return internalServiceContainer.get<IProcessServiceFactory>(IProcessServiceFactory);
@@ -32,8 +29,6 @@ export async function exec(file: string, args: string[], options: SpawnOptions =
     const proc = await getProcessFactory().create();
     return proc.exec(file, args, options);
 }
-
-// filesystem
 
 export function pathExists(absPath: string): Promise<boolean> {
     return fsapi.pathExists(absPath);
@@ -50,19 +45,6 @@ export function readFile(filePath: string): Promise<string> {
  */
 export function isParentPath(filePath: string, parentPath: string): boolean {
     return normCasePath(filePath).startsWith(normCasePath(parentPath));
-}
-
-export function listDir(dirname: string): Promise<string[]> {
-    return fsapi.readdir(dirname);
-}
-
-export async function isDirectory(filename: string): Promise<boolean> {
-    const stat = await fsapi.lstat(filename);
-    return stat.isDirectory();
-}
-
-export function normalizePath(filename: string): string {
-    return normalizeFilename(filename);
 }
 
 export function normCasePath(filePath: string): string {
