+++ src/test/testing/helper.ts
@@ -36,3 +36,43 @@ export function lookForTestFile(tests: Tests, testFile: string) {
 export function fixPath(filename: string): string {
     return filename.replace(/\//, sep);
 }
+
+// Return the indentation part of the given line.
+export function getIndent(line: string): string {
+    const found = line.match(/^ */);
+    return found![0];
+}
+
+// Return the dedented lines in the given text.
+//
+// This is used to represent text concisely and readably, which is
+// particularly useful for declarative definitions (e.g. in tests).
+//
+// (inspired by Python's textwrap.dedent())
+export function getDedentedLines(text: string): string[] {
+    const linesep = text.includes('\r') ? '\r\n' : '\n';
+    const lines = text.split(linesep);
+    if (!lines) {
+        return [text];
+    }
+
+    if (lines[0] !== '') {
+        throw Error('expected actual first line to be blank');
+    }
+    lines.shift();
+
+    if (lines[0] === '') {
+        throw Error('expected "first" line to not be blank');
+    }
+    const leading = getIndent(lines[0]).length;
+
+    for (let i = 0; i < lines.length; i += 1) {
+        const line = lines[i];
+        if (getIndent(line).length < leading) {
+            throw Error(`line ${i} has less indent than the "first" line`);
+        }
+        lines[i] = line.substring(leading);
+    }
+
+    return lines;
+}
